perf(MarketBoard): sort and limit in CQL and use automatic paging (#1018)
Using FetchPageAsync is only recommended if you are going to use a
paging token system. Just allow the DataStax driver to paginate using
the default paging size. This reduced the amount of object allocations
and CPU usage. Additionally, the latency time is faster on my test
machine that checking the cache is now the bottleneck.

Additionally, return IEnumberable directly to allow the caller to be
able to use it more efficiently.

The explicit ORDER BY is to show how we want the data, the table is
already sorted by the sale_time.