update: modify listings from upsert model to replace model.
MVCC allows us to have a transaction view concurrently modified for both
views (grabbing the stall content and inserting the fresh content). This
should prevent any issues with needing a post cleanup job; however, this
will cause use to have VACUUM and page fragmentation, so we should test
this more.